Add path finding plataforms
1) Redo platform code to something that actually works.
2) Add filter to pathfinding system
3) Fix jump code to the pathfinding system
4) Add flying agents pathfinding passable function
5) Add overestimated heuristics formula
6) Tune the heuristic formula for our use case.
7) Fix SEVERAL other bugs in pathfinding system
8) Add agent movement prorpieties.

+++ Assets/Source/Agents/AgentCreationApi.cs
@@ -86,7 +86,7 @@ namespace Agent
             PropertiesArray[CurrentIndex].Name = name;
         }
 
-        public void SetMovement(int defaultSpeed, float jumHegiht, int numOfJumps)
+        public void SetMovement(float defaultSpeed, float jumHegiht, int numOfJumps)
         {
             if (CurrentIndex >= 0 && CurrentIndex < PropertiesArray.Length)
             {
@@ -97,7 +97,7 @@ namespace Agent
             }
         }
 
-        public void SetFlyingMovement(int defaultSpeed)
+        public void SetFlyingMovement(float defaultSpeed)
         {
             if (CurrentIndex >= 0 && CurrentIndex < PropertiesArray.Length)
             {
